temporary workaround: add linkage for some selected CPROVER built-ins
In particular, __new_array is required by the irep cpp_new_array. To have it loaded from CPROVER builtins, we
need to add an external decl to the symtab. We do this similar than in the C(++) parser: add decls in the raw
input (i.e., in JSON), then parse, and finally handover the parser result to the symtab builder